+++ rollup.config.js
@@ -4,5 +4,5 @@ export default {
     file: "index.cjs",
     format: "cjs",
   },
-  external: ["fs", "os", "path", "path/win32", "which"],
+  external: ["fs", "os", "path", "process", "path/win32", "which"],
 };