+++ test/unix.test.js
@@ -479,6 +479,48 @@ describe("unix.js", function () {
                 assert.strictEqual(output, "foobar=\\~:");
               });
 
+              it("escapes a tilde right after '=' with a trailing equals", function () {
+                const input = "foobar=~=";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foobar=\\~=");
+              });
+
+              it("escapes a tilde right after '=' with a trailing plus", function () {
+                const input = "foobar=~-";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foobar=\\~-");
+              });
+
+              it("escapes a tilde right after '=' with a trailing plus", function () {
+                const input = "foobar=~+";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foobar=\\~+");
+              });
+
+              it("escapes a tilde right after '=' with a trailing forward slash", function () {
+                const input = "foobar=~/";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foobar=\\~/");
+              });
+
+              it("escapes a tilde right after '=' with a trailing zero", function () {
+                const input = "foobar=~0";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foobar=\\~0");
+              });
+
+              it("escapes a tilde right after '=' with a trailing whitespace", function () {
+                const input = "foobar=~ ";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foobar=\\~ ");
+              });
+
+              it("escapes a tilde right after '=' with a trailing colon and text", function () {
+                const input = "foo=~:bar";
+                const output = escapeShellArg(input, interpolation);
+                assert.strictEqual(output, "foo=\\~:bar");
+              });
+
               it("escapes a tilde right after '=' with a trailing colon and text", function () {
                 const input = "foo=~:bar";
                 const output = escapeShellArg(input, interpolation);